@@ -1,1 +1,2 @@
-module.exports = require('./lib/mochaReporter');
+export { ReportportalParallelRunAgent } from './src/reportportal-parallel-run-agent'
+export { ReportportalReporter } from './src/reportportal-reporter'
