@@ -14,107 +14,109 @@
  *  limitations under the License.
  */
 
-const EventEmitter = require('events');
-const { getDefaultConfig, RPClient, mockedDate } = require('./mocks');
-const ReportportalAgent = require('./../lib/mochaReporter');
+const EventEmitter = require('events')
+const { getDefaultConfig, RPClient, mockedDate } = require('./mocks')
+const ReportportalAgent = require('./../lib/mochaReporter')
 
-jest.mock('./../lib/utils');
+jest.mock('./../lib/utils')
 
 describe('suites reporting', function () {
-  let reporter;
-  let suiteFirstLevel;
-  let suiteSecondLevel;
-  let rootSuite;
+  let reporter
+  let suiteFirstLevel
+  let suiteSecondLevel
+  let rootSuite
 
   beforeAll(function () {
-    const options = getDefaultConfig();
-    const runner = new EventEmitter();
-    reporter = new ReportportalAgent(runner, options);
-    reporter.rpClient = new RPClient(options);
-    reporter.launchId = 'tempLaunchId';
+    const options = getDefaultConfig()
+    const runner = new EventEmitter()
+    reporter = new ReportportalAgent(runner, options)
+    reporter.rpClient = new RPClient(options)
+    reporter.launchId = 'tempLaunchId'
     rootSuite = {
       title: '',
       root: true,
-    };
+    }
     suiteFirstLevel = {
       title: 'First level suite',
       parent: rootSuite,
-    };
+    }
     suiteSecondLevel = {
       title: 'Second level suite',
       parent: suiteFirstLevel,
-    };
-  });
+    }
+  })
 
   afterEach(function () {
-    reporter.suitesInfo.clear();
-    jest.clearAllMocks();
-  });
+    reporter.suitesInfo.clear()
+    jest.clearAllMocks()
+  })
 
   describe('onSuiteStart', function () {
     it("shouldn't start root suite", function () {
-      const spyStartTestItem = jest.spyOn(reporter.rpClient, 'startTestItem');
+      const spyStartTestItem = jest.spyOn(reporter.rpClient, 'startTestItem')
 
-      reporter.onSuiteStart(rootSuite);
+      reporter.onSuiteStart(rootSuite)
 
-      expect(spyStartTestItem).not.toHaveBeenCalled();
-    });
+      expect(spyStartTestItem).not.toHaveBeenCalled()
+    })
 
     it('should start first-level suite', function () {
-      const spyStartTestItem = jest.spyOn(reporter.rpClient, 'startTestItem');
-      reporter.suitesInfo.set(rootSuite, undefined);
+      const spyStartTestItem = jest.spyOn(reporter.rpClient, 'startTestItem')
+      reporter.suitesInfo.set(rootSuite, undefined)
       const expectedSuiteStartObj = {
         name: 'First level suite',
         startTime: mockedDate,
         attributes: [],
         type: 'suite',
-      };
+      }
 
-      reporter.onSuiteStart(suiteFirstLevel);
+      reporter.onSuiteStart(suiteFirstLevel)
 
       expect(spyStartTestItem).toHaveBeenCalledWith(
         expectedSuiteStartObj,
         'tempLaunchId',
-        undefined,
-      );
-    });
+        undefined
+      )
+    })
 
     it('should start nested suite', function () {
-      const spyStartTestItem = jest.spyOn(reporter.rpClient, 'startTestItem');
-      reporter.suitesInfo.set(rootSuite, undefined);
+      const spyStartTestItem = jest.spyOn(reporter.rpClient, 'startTestItem')
+      reporter.suitesInfo.set(rootSuite, undefined)
       reporter.suitesInfo.set(suiteFirstLevel, {
         tempId: 'firstLevelSuiteId',
         startTime: mockedDate,
-      });
+      })
       const expectedSuiteStartObj = {
         name: 'Second level suite',
         startTime: mockedDate,
         attributes: [],
         type: 'suite',
-      };
+      }
 
-      reporter.onSuiteStart(suiteSecondLevel);
+      reporter.onSuiteStart(suiteSecondLevel)
 
       expect(spyStartTestItem).toHaveBeenCalledWith(
         expectedSuiteStartObj,
         'tempLaunchId',
-        'firstLevelSuiteId',
-      );
-    });
-  });
+        'firstLevelSuiteId'
+      )
+    })
+  })
 
   describe('finishTestItem', function () {
     it('should finish started suite', function () {
-      const spyFinishTestItem = jest.spyOn(reporter.rpClient, 'finishTestItem');
-      reporter.suitesInfo.set(rootSuite, undefined);
+      const spyFinishTestItem = jest.spyOn(reporter.rpClient, 'finishTestItem')
+      reporter.suitesInfo.set(rootSuite, undefined)
       reporter.suitesInfo.set(suiteFirstLevel, {
         tempId: 'firstLevelSuiteId',
         startTime: mockedDate,
-      });
+      })
 
-      reporter.onSuiteFinish(suiteFirstLevel);
+      reporter.onSuiteFinish(suiteFirstLevel)
 
-      expect(spyFinishTestItem).toHaveBeenCalledWith('firstLevelSuiteId', { endTime: mockedDate });
-    });
-  });
-});
+      expect(spyFinishTestItem).toHaveBeenCalledWith('firstLevelSuiteId', {
+        endTime: mockedDate,
+      })
+    })
+  })
+})
