@@ -14,93 +14,102 @@
  *  limitations under the License.
  */
 
-const EventEmitter = require('events');
-const { getDefaultConfig, RPClient, mockedDate } = require('./mocks');
-const ReportportalAgent = require('./../lib/mochaReporter');
-const testStatuses = require('./../lib/constants/testStatuses');
+const EventEmitter = require('events')
+const { getDefaultConfig, RPClient, mockedDate } = require('./mocks')
+const ReportportalAgent = require('./../lib/mochaReporter')
+const testStatuses = require('./../lib/constants/testStatuses')
 
-jest.mock('./../lib/utils');
+jest.mock('./../lib/utils')
 
 describe('test items reporting', function () {
-  let reporter;
-  let suite;
+  let reporter
+  let suite
 
   const createAndPrepareReporter = (customReporterOptions = {}) => {
-    const options = getDefaultConfig();
-    Object.assign(options.reporterOptions, customReporterOptions);
-    const runner = new EventEmitter();
-    reporter = new ReportportalAgent(runner, options);
-    reporter.rpClient = new RPClient(options);
-    reporter.launchId = 'tempLaunchId';
+    const options = getDefaultConfig()
+    Object.assign(options.reporterOptions, customReporterOptions)
+    const runner = new EventEmitter()
+    reporter = new ReportportalAgent(runner, options)
+    reporter.rpClient = new RPClient(options)
+    reporter.launchId = 'tempLaunchId'
     suite = {
       title: 'Suite',
       parent: {},
-    };
-    reporter.suitesInfo.set(suite, { tempId: 'tempSuiteId', startTime: mockedDate });
-    return reporter;
-  };
+    }
+    reporter.suitesInfo.set(suite, {
+      tempId: 'tempSuiteId',
+      startTime: mockedDate,
+    })
+    return reporter
+  }
 
   describe('finishTest', function () {
     afterEach(function () {
-      reporter.currentTest = null;
-      reporter.hookIds.clear();
-      reporter.attributes.clear();
-      reporter.descriptions.clear();
-      jest.clearAllMocks();
-    });
+      reporter.currentTest = null
+      reporter.hookIds.clear()
+      reporter.attributes.clear()
+      reporter.descriptions.clear()
+      jest.clearAllMocks()
+    })
     it('should finish test with specified status', function () {
-      reporter = createAndPrepareReporter();
-      const spyFinishTestItem = jest.spyOn(reporter.rpClient, 'finishTestItem');
+      reporter = createAndPrepareReporter()
+      const spyFinishTestItem = jest.spyOn(reporter.rpClient, 'finishTestItem')
       const currentTest = {
         title: 'test',
         parent: suite,
         state: 'passed',
         tempId: 'testItemId',
-      };
+      }
       const expectedTestFinishObj = {
         endTime: mockedDate,
         status: 'failed',
         retry: false,
-      };
-      reporter.currentTest = currentTest;
+      }
+      reporter.currentTest = currentTest
 
-      reporter.finishTest(currentTest, testStatuses.FAILED);
+      reporter.finishTest(currentTest, testStatuses.FAILED)
 
-      expect(spyFinishTestItem).toHaveBeenCalledWith('testItemId', expectedTestFinishObj);
-    });
+      expect(spyFinishTestItem).toHaveBeenCalledWith(
+        'testItemId',
+        expectedTestFinishObj
+      )
+    })
 
     it('skippedIssue is not defined: should finish skipped test with issue', function () {
-      reporter = createAndPrepareReporter();
-      const spyFinishTestItem = jest.spyOn(reporter.rpClient, 'finishTestItem');
+      reporter = createAndPrepareReporter()
+      const spyFinishTestItem = jest.spyOn(reporter.rpClient, 'finishTestItem')
       const currentTest = {
         title: 'test',
         parent: suite,
         state: 'pending',
         tempId: 'testItemId',
-      };
+      }
       const expectedTestFinishObj = {
         endTime: mockedDate,
         status: 'skipped',
         retry: false,
-      };
-      reporter.currentTest = currentTest;
+      }
+      reporter.currentTest = currentTest
 
-      reporter.finishTest(currentTest, testStatuses.SKIPPED);
+      reporter.finishTest(currentTest, testStatuses.SKIPPED)
 
-      expect(spyFinishTestItem).toHaveBeenCalledWith('testItemId', expectedTestFinishObj);
-    });
+      expect(spyFinishTestItem).toHaveBeenCalledWith(
+        'testItemId',
+        expectedTestFinishObj
+      )
+    })
 
     it('skippedIssue=false: should finish skipped test with issue NOT_ISSUE', function () {
       reporter = createAndPrepareReporter({
         skippedIssue: false,
-      });
-      const spyFinishTestItem = jest.spyOn(reporter.rpClient, 'finishTestItem');
+      })
+      const spyFinishTestItem = jest.spyOn(reporter.rpClient, 'finishTestItem')
       const currentTest = {
         title: 'test',
         parent: suite,
         state: 'pending',
         tempId: 'testItemId',
-      };
+      }
       const expectedTestFinishObj = {
         endTime: mockedDate,
         status: 'skipped',
@@ -108,314 +117,362 @@ describe('test items reporting', function () {
         issue: {
           issueType: 'NOT_ISSUE',
         },
-      };
-      reporter.currentTest = currentTest;
+      }
+      reporter.currentTest = currentTest
 
-      reporter.finishTest(currentTest, testStatuses.SKIPPED);
+      reporter.finishTest(currentTest, testStatuses.SKIPPED)
 
-      expect(spyFinishTestItem).toHaveBeenCalledWith('testItemId', expectedTestFinishObj);
-    });
+      expect(spyFinishTestItem).toHaveBeenCalledWith(
+        'testItemId',
+        expectedTestFinishObj
+      )
+    })
 
     it('attributes exists for test: should finish test with corresponded attributes', function () {
-      reporter = createAndPrepareReporter();
-      const spyFinishTestItem = jest.spyOn(reporter.rpClient, 'finishTestItem');
+      reporter = createAndPrepareReporter()
+      const spyFinishTestItem = jest.spyOn(reporter.rpClient, 'finishTestItem')
       const currentTest = {
         tempId: 'testItemId',
-      };
-      reporter.attributes.set('testItemId', [{ key: 'key1', value: 'value1' }]);
+      }
+      reporter.attributes.set('testItemId', [{ key: 'key1', value: 'value1' }])
 
       const expectedTestFinishObj = {
         endTime: mockedDate,
         retry: false,
         status: 'passed',
         attributes: [{ key: 'key1', value: 'value1' }],
-      };
-      reporter.currentTest = currentTest;
+      }
+      reporter.currentTest = currentTest
 
-      reporter.finishTest(currentTest, testStatuses.PASSED);
+      reporter.finishTest(currentTest, testStatuses.PASSED)
 
-      expect(spyFinishTestItem).toHaveBeenCalledWith('testItemId', expectedTestFinishObj);
-    });
+      expect(spyFinishTestItem).toHaveBeenCalledWith(
+        'testItemId',
+        expectedTestFinishObj
+      )
+    })
     it('description exists for the test: should finish test with description', function () {
-      reporter = createAndPrepareReporter();
-      const spyFinishTestItem = jest.spyOn(reporter.rpClient, 'finishTestItem');
+      reporter = createAndPrepareReporter()
+      const spyFinishTestItem = jest.spyOn(reporter.rpClient, 'finishTestItem')
       const currentTest = {
         tempId: 'testItemId',
-      };
-      reporter.descriptions.set('testItemId', 'test description');
+      }
+      reporter.descriptions.set('testItemId', 'test description')
 
       const expectedTestFinishObj = {
         endTime: mockedDate,
         retry: false,
         status: 'passed',
         description: 'test description',
-      };
-      reporter.currentTest = currentTest;
+      }
+      reporter.currentTest = currentTest
 
-      reporter.finishTest(currentTest, testStatuses.PASSED);
+      reporter.finishTest(currentTest, testStatuses.PASSED)
 
-      expect(spyFinishTestItem).toHaveBeenCalledWith('testItemId', expectedTestFinishObj);
-    });
+      expect(spyFinishTestItem).toHaveBeenCalledWith(
+        'testItemId',
+        expectedTestFinishObj
+      )
+    })
     it('testCaseId exists for the test: should finish test with testCaseId', function () {
-      reporter = createAndPrepareReporter();
-      const spyFinishTestItem = jest.spyOn(reporter.rpClient, 'finishTestItem');
+      reporter = createAndPrepareReporter()
+      const spyFinishTestItem = jest.spyOn(reporter.rpClient, 'finishTestItem')
       const currentTest = {
         tempId: 'testItemId',
-      };
-      reporter.testCaseIds.set('testItemId', 'test_case_Id');
+      }
+      reporter.testCaseIds.set('testItemId', 'test_case_Id')
 
       const expectedTestFinishObj = {
         endTime: mockedDate,
         retry: false,
         status: 'passed',
         testCaseId: 'test_case_Id',
-      };
-      reporter.currentTest = currentTest;
+      }
+      reporter.currentTest = currentTest
 
-      reporter.finishTest(currentTest, testStatuses.PASSED);
+      reporter.finishTest(currentTest, testStatuses.PASSED)
 
-      expect(spyFinishTestItem).toHaveBeenCalledWith('testItemId', expectedTestFinishObj);
-    });
+      expect(spyFinishTestItem).toHaveBeenCalledWith(
+        'testItemId',
+        expectedTestFinishObj
+      )
+    })
 
     it('custom status assigned for the test: should finish test with custom status', function () {
-      reporter = createAndPrepareReporter();
-      const spyFinishTestItem = jest.spyOn(reporter.rpClient, 'finishTestItem');
+      reporter = createAndPrepareReporter()
+      const spyFinishTestItem = jest.spyOn(reporter.rpClient, 'finishTestItem')
       const currentTest = {
         tempId: 'testItemId',
-      };
+      }
 
       const expectedTestFinishObj = {
         endTime: mockedDate,
         retry: false,
         status: 'info',
-      };
-      reporter.currentTest = currentTest;
-      reporter.setStatus({ status: 'info' });
+      }
+      reporter.currentTest = currentTest
+      reporter.setStatus({ status: 'info' })
 
-      reporter.finishTest(currentTest, testStatuses.PASSED);
+      reporter.finishTest(currentTest, testStatuses.PASSED)
 
-      expect(spyFinishTestItem).toHaveBeenCalledWith('testItemId', expectedTestFinishObj);
-    });
-  });
+      expect(spyFinishTestItem).toHaveBeenCalledWith(
+        'testItemId',
+        expectedTestFinishObj
+      )
+    })
+  })
 
   describe('test event listeners', function () {
     beforeAll(function () {
-      reporter = createAndPrepareReporter();
-    });
+      reporter = createAndPrepareReporter()
+    })
 
     afterEach(function () {
-      reporter.currentTest = null;
-      reporter.hookIds.clear();
-      jest.clearAllMocks();
-    });
+      reporter.currentTest = null
+      reporter.hookIds.clear()
+      jest.clearAllMocks()
+    })
 
     describe('onTestStart', function () {
       it('should start test', function () {
-        const spyStartTestItem = jest.spyOn(reporter.rpClient, 'startTestItem');
+        const spyStartTestItem = jest.spyOn(reporter.rpClient, 'startTestItem')
         const testItem = {
           title: 'test',
           parent: suite,
-        };
+        }
         const expectedTestStartObj = {
           name: 'test',
           startTime: mockedDate,
           attributes: [],
           type: 'step',
           retry: false,
-        };
+        }
 
-        reporter.onTestStart(testItem);
+        reporter.onTestStart(testItem)
 
         expect(spyStartTestItem).toHaveBeenCalledWith(
           expectedTestStartObj,
           'tempLaunchId',
-          'tempSuiteId',
-        );
-      });
+          'tempSuiteId'
+        )
+      })
 
       it('should start first retry', function () {
-        const spyStartTestItem = jest.spyOn(reporter.rpClient, 'startTestItem');
+        const spyStartTestItem = jest.spyOn(reporter.rpClient, 'startTestItem')
         const testItem = {
           title: 'test',
           parent: suite,
           _retries: 2,
-        };
+        }
         const expectedTestStartObj = {
           name: 'test',
           startTime: mockedDate,
           attributes: [],
           type: 'step',
           retry: true,
-        };
+        }
 
-        reporter.onTestStart(testItem);
+        reporter.onTestStart(testItem)
 
         expect(spyStartTestItem).toHaveBeenCalledWith(
           expectedTestStartObj,
           'tempLaunchId',
-          'tempSuiteId',
-        );
-      });
+          'tempSuiteId'
+        )
+      })
 
       it('should finish first and start second retry', function () {
-        const spyStartTestItem = jest.spyOn(reporter.rpClient, 'startTestItem');
-        const spyFinishTestItem = jest.spyOn(reporter.rpClient, 'finishTestItem');
+        const spyStartTestItem = jest.spyOn(reporter.rpClient, 'startTestItem')
+        const spyFinishTestItem = jest.spyOn(
+          reporter.rpClient,
+          'finishTestItem'
+        )
         const testItem = {
           title: 'test',
           parent: suite,
           _retries: 2,
-        };
+        }
         const expectedTestStartObj = {
           name: 'test',
           startTime: mockedDate,
           attributes: [],
           type: 'step',
           retry: true,
-        };
+        }
         const expectedTestFinishObj = {
           status: 'failed',
           endTime: mockedDate,
           retry: true,
-        };
+        }
 
-        reporter.onTestStart(testItem);
-        reporter.onTestStart(testItem);
+        reporter.onTestStart(testItem)
+        reporter.onTestStart(testItem)
 
-        expect(spyStartTestItem).toHaveBeenCalledTimes(2);
+        expect(spyStartTestItem).toHaveBeenCalledTimes(2)
         expect(spyStartTestItem).toHaveBeenCalledWith(
           expectedTestStartObj,
           'tempLaunchId',
-          'tempSuiteId',
-        );
-        expect(spyFinishTestItem).toHaveBeenCalledWith('testItemId', expectedTestFinishObj);
-      });
-    });
+          'tempSuiteId'
+        )
+        expect(spyFinishTestItem).toHaveBeenCalledWith(
+          'testItemId',
+          expectedTestFinishObj
+        )
+      })
+    })
 
     describe('onTestPending', function () {
       it('should start and finish pending test', function () {
-        const spyStartTestItem = jest.spyOn(reporter.rpClient, 'startTestItem');
-        const spyFinishTestItem = jest.spyOn(reporter.rpClient, 'finishTestItem');
+        const spyStartTestItem = jest.spyOn(reporter.rpClient, 'startTestItem')
+        const spyFinishTestItem = jest.spyOn(
+          reporter.rpClient,
+          'finishTestItem'
+        )
         const testItem = {
           title: 'test',
           parent: suite,
           state: 'pending',
-        };
+        }
         const expectedTestStartObj = {
           name: 'test',
           startTime: mockedDate,
           attributes: [],
           type: 'step',
           retry: false,
-        };
+        }
         const expectedTestFinishObj = {
           endTime: mockedDate,
           status: 'skipped',
           retry: false,
-        };
+        }
 
-        reporter.onTestPending(testItem);
+        reporter.onTestPending(testItem)
 
         expect(spyStartTestItem).toHaveBeenCalledWith(
           expectedTestStartObj,
           'tempLaunchId',
-          'tempSuiteId',
-        );
-        expect(spyFinishTestItem).toHaveBeenCalledWith('testItemId', expectedTestFinishObj);
-      });
-    });
+          'tempSuiteId'
+        )
+        expect(spyFinishTestItem).toHaveBeenCalledWith(
+          'testItemId',
+          expectedTestFinishObj
+        )
+      })
+    })
 
     describe('onTestFinish', function () {
       it('should finish passed test', function () {
-        const spyFinishTestItem = jest.spyOn(reporter.rpClient, 'finishTestItem');
+        const spyFinishTestItem = jest.spyOn(
+          reporter.rpClient,
+          'finishTestItem'
+        )
         const currentTest = {
           title: 'test',
           parent: suite,
           state: 'passed',
           tempId: 'testItemId',
-        };
+        }
         const expectedTestFinishObj = {
           endTime: mockedDate,
           status: 'passed',
           retry: false,
-        };
-        reporter.currentTest = currentTest;
+        }
+        reporter.currentTest = currentTest
 
-        reporter.onTestFinish(currentTest);
+        reporter.onTestFinish(currentTest)
 
-        expect(spyFinishTestItem).toHaveBeenCalledWith('testItemId', expectedTestFinishObj);
-      });
+        expect(spyFinishTestItem).toHaveBeenCalledWith(
+          'testItemId',
+          expectedTestFinishObj
+        )
+      })
 
       it('should finish failed test', function () {
-        const spyFinishTestItem = jest.spyOn(reporter.rpClient, 'finishTestItem');
+        const spyFinishTestItem = jest.spyOn(
+          reporter.rpClient,
+          'finishTestItem'
+        )
         const currentTest = {
           title: 'test',
           parent: suite,
           state: 'failed',
           tempId: 'testItemId',
-        };
+        }
         const expectedTestFinishObj = {
           endTime: mockedDate,
           status: 'failed',
           retry: false,
-        };
-        reporter.currentTest = currentTest;
+        }
+        reporter.currentTest = currentTest
 
-        reporter.onTestFinish(currentTest);
+        reporter.onTestFinish(currentTest)
 
-        expect(spyFinishTestItem).toHaveBeenCalledWith('testItemId', expectedTestFinishObj);
-      });
+        expect(spyFinishTestItem).toHaveBeenCalledWith(
+          'testItemId',
+          expectedTestFinishObj
+        )
+      })
 
       it('should finish failed retry', function () {
-        const spyFinishTestItem = jest.spyOn(reporter.rpClient, 'finishTestItem');
+        const spyFinishTestItem = jest.spyOn(
+          reporter.rpClient,
+          'finishTestItem'
+        )
         const currentTest = {
           title: 'test',
           parent: suite,
           state: 'failed',
           _retries: 2,
           tempId: 'testItemId',
-        };
+        }
         const expectedTestFinishObj = {
           endTime: mockedDate,
           status: 'failed',
           retry: true,
-        };
-        reporter.currentTest = currentTest;
+        }
+        reporter.currentTest = currentTest
 
-        reporter.onTestFinish(currentTest);
+        reporter.onTestFinish(currentTest)
 
-        expect(spyFinishTestItem).toHaveBeenCalledWith('testItemId', expectedTestFinishObj);
-      });
-    });
+        expect(spyFinishTestItem).toHaveBeenCalledWith(
+          'testItemId',
+          expectedTestFinishObj
+        )
+      })
+    })
 
     describe('onTestFail', function () {
       it('should send log on test fail', function () {
-        const spySendLog = jest.spyOn(reporter.rpClient, 'sendLog');
+        const spySendLog = jest.spyOn(reporter.rpClient, 'sendLog')
         const currentTest = {
           title: 'test',
           parent: suite,
           state: 'failed',
           tempId: 'testItemId',
-        };
+        }
         const expectedLogObj = {
           level: 'ERROR',
           message: 'error message',
-        };
-        reporter.currentTest = currentTest;
+        }
+        reporter.currentTest = currentTest
 
-        reporter.onTestFail(currentTest, 'error message');
+        reporter.onTestFail(currentTest, 'error message')
 
-        expect(spySendLog).toHaveBeenCalledWith('testItemId', expectedLogObj);
-      });
+        expect(spySendLog).toHaveBeenCalledWith('testItemId', expectedLogObj)
+      })
 
       it('should finish current test as skipped on hook fail', function () {
-        const spyFinishTestItem = jest.spyOn(reporter.rpClient, 'finishTestItem');
+        const spyFinishTestItem = jest.spyOn(
+          reporter.rpClient,
+          'finishTestItem'
+        )
         const currentTest = {
           title: 'test #1',
           parent: suite,
           state: 'pending',
           tempId: 'testItemId',
-        };
-        reporter.currentTest = currentTest;
+        }
+        reporter.currentTest = currentTest
         const hook = {
           title: '"before each" hook: named hook',
           parent: suite,
@@ -424,20 +481,23 @@ describe('test items reporting', function () {
           ctx: {
             currentTest,
           },
-        };
+        }
         const expectedTestFinishObj = {
           endTime: mockedDate,
           status: 'skipped',
           retry: false,
-        };
-        reporter.hookIds.set(hook, 'hookId');
-        hook.state = 'failed';
+        }
+        reporter.hookIds.set(hook, 'hookId')
+        hook.state = 'failed'
 
-        reporter.onTestFail(hook, 'error message');
+        reporter.onTestFail(hook, 'error message')
 
-        expect(spyFinishTestItem).toHaveBeenCalledTimes(1);
-        expect(spyFinishTestItem).toHaveBeenCalledWith('testItemId', expectedTestFinishObj);
-      });
-    });
-  });
-});
+        expect(spyFinishTestItem).toHaveBeenCalledTimes(1)
+        expect(spyFinishTestItem).toHaveBeenCalledWith(
+          'testItemId',
+          expectedTestFinishObj
+        )
+      })
+    })
+  })
+})
